# gen/http/openapi3.json
{"openapi":"3.0.3","info":{"title":"LFX V2 - Access Check Service","description":"LFX Access Check Service for bulk access checks","version":"0.0.1"},"servers":[{"url":"http://localhost:80","description":"Default server for access-svc"}],"paths":{"/access-check":{"post":{"tags":["access-svc"],"summary":"check-access access-svc","description":"Check access permissions for resource-action pairs","operationId":"access-svc#check-access","parameters":[{"name":"v","in":"query","description":"API version","allowEmptyValue":true,"required":true,"schema":{"type":"string","description":"API version","example":"1","enum":["1"]},"example":"1"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CheckAccessRequestBody"},"example":{"requests":["project:123#read","committee:456#write"]}}}},"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CheckAccessResponseBody"},"example":{"results":["allow","deny"]}}}},"400":{"description":"BadRequest: Bad request","content":{"application/vnd.goa.error":{"schema":{"$ref":"#/components/schemas/Error"}}}},"401":{"description":"Unauthorized: Unauthorized","content":{"application/vnd.goa.error":{"schema":{"$ref":"#/components/schemas/Error"}}}}},"security":[{"jwt_header_Authorization":[]}]}},"/my-grants":{"get":{"tags":["access-svc"],"summary":"my-grants access-svc","description":"Get the caller's direct access grants for a given object type","operationId":"access-svc#my-grants","parameters":[{"name":"v","in":"query","description":"API version","allowEmptyValue":true,"required":true,"schema":{"type":"string","description":"API version","example":"1","enum":["1"]},"example":"1"},{"name":"object_type","in":"query","description":"Object type to query grants for","allowEmptyValue":true,"required":true,"schema":{"type":"string","description":"Object type to query grants for","example":"project","pattern":"^[a-z]+(_[a-z]+)*$"},"example":"project"}],"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MyGrantsResponseBody"},"example":{"grants":["project:a27394a3-7a6c-4d0f-9e0f-692d8753924f#writer@user:auth0|bramwelt"]}}}},"400":{"description":"BadRequest: Bad request","content":{"application/vnd.goa.error":{"schema":{"$ref":"#/components/schemas/Error"}}}},"401":{"description":"Unauthorized: Unauthorized","content":{"application/vnd.goa.error":{"schema":{"$ref":"#/components/schemas/Error"}}}}},"security":[{"jwt_header_Authorization":[]}]}}},"components":{"schemas":{"AccessErrorResult":{"type":"object","properties":{"code":{"type":"string","description":"Error code","example":"INVALID_REQUEST"},"message":{"type":"string","description":"Error message","example":"Invalid request format"}},"description":"Standard error response for access check service","example":{"code":"INVALID_REQUEST","message":"Invalid request format"},"required":["message"]},"CheckAccessRequestBody":{"type":"object","properties":{"requests":{"type":"array","items":{"type":"string","example":"Quos fugiat praesentium aut aut."},"description":"Resource-action pairs to check","example":["project:123#read","committee:456#write"],"minItems":1}},"example":{"requests":["project:123#read","committee:456#write"]},"required":["requests"]},"CheckAccessResponseBody":{"type":"object","properties":{"results":{"type":"array","items":{"type":"string","example":"Dolores voluptas occaecati."},"description":"Access check results","example":["allow","deny"]}},"example":{"results":["allow","deny"]},"required":["results"]},"Error":{"type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":true},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":false},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":true}},"description":"Bad request","example":{"fault":true,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":false,"timeout":true},"required":["name","id","message","temporary","timeout","fault"]},"MyGrantsResponseBody":{"type":"object","properties":{"grants":{"type":"array","items":{"type":"string","example":"Sapiente architecto animi."},"description":"Direct access grants as tuple-strings","example":["project:a27394a3-7a6c-4d0f-9e0f-692d8753924f#writer@user:auth0|bramwelt"]}},"example":{"grants":["project:a27394a3-7a6c-4d0f-9e0f-692d8753924f#writer@user:auth0|bramwelt"]},"required":["grants"]}},"securitySchemes":{"jwt_header_Authorization":{"type":"http","description":"Heimdall authorization","scheme":"bearer"}}},"tags":[{"name":"access-svc","description":"LFX Access Check Service"}]}